requirements.txt
absl-py==0.15.0
aiohttp==3.7.4.post0
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.0.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==7.1.2
cryptography==35.0.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
flatbuffers==2.0
gast==0.5.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.17
grpcio==1.41.0
h5py==2.10.0
idna==3.1
importlib-metadata==4.8.1
ipykernel==5.3.4
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
keras-tuner==1.0.4
kiwisolver==1.3.2
kt-legacy==1.0.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.1
numpy==1.21.3
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==20.2.4
pipreqs==0.4.11
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.18.1
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pydeck==0.7.0
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyu2f==0.1.5
pyzmq==20.0.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
setuptools==58.2.0
six==1.16.0
smmap==5.0.0
split-folders==0.4.3
streamlit==1.1.0
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-addons-macos==0.1a3
tensorflow-estimator==2.6.0
tensorflow-macos==0.1a3
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
typeguard==2.13.0
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.7
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
wrapt==1.13.2
yarg==0.1.9
yarl==1.7.0
zipp==3.6.0
